Fix the ImplicitObjectRequired canonicalizer rule
